+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 1. [#302](https://github.com/influxdata/influxdb-client-js/pull/302): Expose response headers in all APIs.
 1. [#302](https://github.com/influxdata/influxdb-client-js/pull/302): Add `tokens.js` example.
 1. [#306](https://github.com/influxdata/influxdb-client-js/pull/306): Regenerate APIs from swagger.
+1. [#309](https://github.com/influxdata/influxdb-client-js/pull/309): Allow to customize logger.
 
 ### Bug Fixes